test(serverless): refactoriza los tests de la libreria comun
- corrige los imports rotos de los tests de shared: apuntaban a los
  modulos planos viejos (shared.ulid, shared.config, shared.cors, ...)
  cuando shared se reorganizo en subpaquetes (shared.core.ulid,
  shared.core.config, shared.http.cors, ...)
- reorganiza los 9 tests planos de shared/tests/unit/shared/ en subdirs
  por subpaquete (core/, aws/, http/) para que --shared=<subpaquete>
  funcione uniformemente en los 7 subpaquetes con tests
- el comando tests --type=coverage pasa --cov-config apuntando al
  pyproject del backend: el omit de core/shared/ evita contar la
  libreria vendorizada en el coverage de cada Lambda
- con --shared=<subpaquete> el --cov se acota a ese subpaquete
- los 254 tests de shared y los 70 de los 4 Lambdas pasan
