+++ serverless/lambda/shared/tests/unit/shared/http/test_responses.py
@@ -1,4 +1,4 @@
-"""Unit tests para shared.responses."""
+"""Unit tests para shared.http.responses."""
 
 from __future__ import annotations
 
@@ -6,12 +6,12 @@ import json
 
 import pytest
 
-from shared.exceptions import (
+from shared.core.exceptions import (
     IPBlacklistedError,
     RateLimitExceededError,
     ValidationError,
 )
-from shared.responses import error_response, json_response, no_content_response
+from shared.http.responses import error_response, json_response, no_content_response
 
 pytestmark = pytest.mark.unit
